# contracts/openapi/teamos.yaml
openapi: 3.1.0
...
info:
  title: TeamOS Gateway API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Внешний JSON/REST-контракт TeamOS. Поля и формы ответов зеркалят
    `src/api/index.ts` и `src/types/index.ts` frontend-приложения.
    Все даты-время передаются в ISO-8601, идентификаторы являются UUID.
servers:
  - url: http://localhost:8080
    description: Локальный gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tasks/boards:
    get:
      tags: [Tasks]
      operationId: getBoards
      summary: Получить доски
      responses:
        '200':
          $ref: '#/components/responses/Boards'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/boards/{boardId}/columns:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    get:
      tags: [Tasks]
      operationId: getColumns
      summary: Получить колонки доски
      responses:
        '200':
          $ref: '#/components/responses/TaskColumns'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Tasks]
      operationId: createColumn
      summary: Создать колонку доски
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskColumnInput'
      responses:
        '201':
          $ref: '#/components/responses/TaskColumn'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/columns/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Tasks]
      operationId: updateColumn
      summary: Обновить колонку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskColumnInput'
      responses:
        '200':
          $ref: '#/components/responses/TaskColumn'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks:
    get:
      tags: [Tasks]
      operationId: getTasks
      summary: Получить задачи
      parameters:
        - name: boardId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          $ref: '#/components/responses/Tasks'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Создать задачу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskInput'
      responses:
        '201':
          $ref: '#/components/responses/Task'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Получить задачу
      responses:
        '200':
          $ref: '#/components/responses/Task'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: Обновить задачу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskInput'
      responses:
        '200':
          $ref: '#/components/responses/Task'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/{taskId}/move:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    post:
      tags: [Tasks]
      operationId: moveTask
      summary: Переместить задачу и изменить порядок
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTaskInput'
      responses:
        '200':
          $ref: '#/components/responses/Task'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/{taskId}/comments:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    get:
      tags: [Tasks]
      operationId: getComments
      summary: Получить комментарии задачи
      responses:
        '200':
          $ref: '#/components/responses/TaskComments'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Tasks]
      operationId: addComment
      summary: Добавить комментарий
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTaskCommentInput'
      responses:
        '201':
          $ref: '#/components/responses/TaskComment'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tasks/labels:
    get:
      tags: [Tasks]
      operationId: getLabels
      summary: Получить метки задач
      responses:
        '200':
          $ref: '#/components/responses/Labels'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: teamos_refresh
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ID'
...
  responses:
...
    Error:
      description: Ошибка API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Boards:
      description: Доски задач
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Board'
    TaskColumn:
      description: Колонка задач
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TaskColumn'
    TaskColumns:
      description: Колонки задач
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TaskColumn'
    Task:
      description: Задача
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Task'
    Tasks:
      description: Задачи
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Task'
    TaskComment:
      description: Комментарий задачи
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TaskComment'
    TaskComments:
      description: Комментарии задачи
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TaskComment'
    Labels:
      description: Метки задач
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Label'
...
  schemas:
    ID:
      type: string
      format: uuid
    ISODateTime:
      type: string
      format: date-time
...
    ClearableDateTime:
      oneOf:
        - $ref: '#/components/schemas/ISODateTime'
        - type: string
          const: ''
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, status]
          properties:
            message:
              type: string
              minLength: 1
              example: Сотрудник не найден
            status:
              type: integer
              minimum: 400
              maximum: 599
              example: 404
    RichTextContent:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          const: doc
        content:
          type: array
          items: {}
...
    BoardType:
      type: string
      enum: [personal, department, project]
    Board:
      type: object
      additionalProperties: false
      required: [id, name, type, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        type:
          $ref: '#/components/schemas/BoardType'
        departmentId:
          $ref: '#/components/schemas/ID'
        ownerId:
          $ref: '#/components/schemas/ID'
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
    TaskColumn:
      type: object
      additionalProperties: false
      required: [id, boardId, name, order]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        boardId:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        order:
          type: integer
          minimum: 0
        color:
          type: string
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
    Label:
      type: object
      additionalProperties: false
      required: [id, name, color]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        color:
          type: string
    ChecklistItem:
      type: object
      additionalProperties: false
      required: [id, text, done]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        text:
          type: string
        done:
          type: boolean
    Attachment:
      type: object
      additionalProperties: false
      required: [id, name, url, size, mimeType]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        url:
          type: string
          format: uri
        size:
          type: integer
          minimum: 0
        mimeType:
          type: string
    TaskSourceType:
      type: string
      enum: [task, contact, company, deal]
    TaskSource:
      type: object
      additionalProperties: false
      required: [type, title, url]
      properties:
        type:
          $ref: '#/components/schemas/TaskSourceType'
        title:
          type: string
        url:
          type: string
        funnelName:
          type: string
        stageName:
          type: string
    RecurrenceFrequency:
      type: string
      enum: [daily, weekly, monthly]
    RecurrenceRule:
      type: object
      additionalProperties: false
      required: [frequency, interval]
      properties:
        frequency:
          $ref: '#/components/schemas/RecurrenceFrequency'
        interval:
          type: integer
          minimum: 1
        weekdays:
          type: array
          uniqueItems: true
          items:
            type: integer
            minimum: 0
            maximum: 6
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - boardId
        - columnId
        - order
        - title
        - authorId
        - assigneeIds
        - watcherIds
        - priority
        - labelIds
        - checklist
        - attachments
        - linkedArticleIds
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ID'
        boardId:
          $ref: '#/components/schemas/ID'
        columnId:
          $ref: '#/components/schemas/ID'
        order:
          type: integer
          minimum: 0
        title:
          type: string
          minLength: 1
        description:
          $ref: '#/components/schemas/RichTextContent'
        authorId:
          $ref: '#/components/schemas/ID'
        assigneeIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        assigneePositionId:
          $ref: '#/components/schemas/ID'
        watcherIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        dueDate:
          $ref: '#/components/schemas/ISODateTime'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        labelIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        source:
          $ref: '#/components/schemas/TaskSource'
        linkedArticleIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        completedAt:
          $ref: '#/components/schemas/ISODateTime'
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
    TaskComment:
      type: object
      additionalProperties: false
      required: [id, taskId, authorId, content, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        taskId:
          $ref: '#/components/schemas/ID'
        authorId:
          $ref: '#/components/schemas/ID'
        content:
          $ref: '#/components/schemas/RichTextContent'
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
    CreateTaskColumnInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
    UpdateTaskColumnInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
    CreateTaskInput:
      type: object
      additionalProperties: false
      required: [boardId, columnId, title]
      properties:
        boardId:
          $ref: '#/components/schemas/ID'
        columnId:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          minLength: 1
        priority:
          $ref: '#/components/schemas/TaskPriority'
    UpdateTaskInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
        description:
          $ref: '#/components/schemas/RichTextContent'
        assigneeIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        assigneePositionId:
          $ref: '#/components/schemas/ID'
        watcherIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        dueDate:
          $ref: '#/components/schemas/ClearableDateTime'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        labelIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        linkedArticleIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        completedAt:
          $ref: '#/components/schemas/ClearableDateTime'
    MoveTaskInput:
      type: object
      additionalProperties: false
      required: [columnId, order]
      properties:
        columnId:
          $ref: '#/components/schemas/ID'
        order:
          type: integer
          minimum: 0
    AddTaskCommentInput:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          $ref: '#/components/schemas/RichTextContent'
...