# services/tasks/internal/storage/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Board struct {
	ID           uuid.UUID     `json:"id"`
	CompanyID    uuid.UUID     `json:"company_id"`
	Name         string        `json:"name"`
	Type         string        `json:"type"`
	DepartmentID uuid.NullUUID `json:"department_id"`
	OwnerID      uuid.NullUUID `json:"owner_id"`
	CreatedAt    time.Time     `json:"created_at"`
}

type Column struct {
	ID      uuid.UUID   `json:"id"`
	BoardID uuid.UUID   `json:"board_id"`
	Name    string      `json:"name"`
	Color   pgtype.Text `json:"color"`
	Order   int32       `json:"order"`
}

type Comment struct {
	ID        uuid.UUID `json:"id"`
	TaskID    uuid.UUID `json:"task_id"`
	AuthorID  uuid.UUID `json:"author_id"`
	Content   []byte    `json:"content"`
	CreatedAt time.Time `json:"created_at"`
}

type Label struct {
	ID        uuid.UUID `json:"id"`
	CompanyID uuid.UUID `json:"company_id"`
	Name      string    `json:"name"`
	Color     string    `json:"color"`
}

type Outbox struct {
	ID            uuid.UUID          `json:"id"`
	CompanyID     uuid.UUID          `json:"company_id"`
	AggregateID   uuid.UUID          `json:"aggregate_id"`
	EventOrder    int64              `json:"event_order"`
	Subject       string             `json:"subject"`
	Payload       []byte             `json:"payload"`
	Headers       []byte             `json:"headers"`
	OccurredAt    time.Time          `json:"occurred_at"`
	NextAttemptAt time.Time          `json:"next_attempt_at"`
	PublishedAt   pgtype.Timestamptz `json:"published_at"`
	Attempts      int32              `json:"attempts"`
	LastError     pgtype.Text        `json:"last_error"`
}

type ProcessedEvent struct {
	EventID     uuid.UUID `json:"event_id"`
	CompanyID   uuid.UUID `json:"company_id"`
	ProcessedAt time.Time `json:"processed_at"`
}

type Task struct {
	ID                    uuid.UUID          `json:"id"`
	CompanyID             uuid.UUID          `json:"company_id"`
	BoardID               uuid.UUID          `json:"board_id"`
	ColumnID              uuid.UUID          `json:"column_id"`
	Order                 int32              `json:"order"`
	Title                 string             `json:"title"`
	Description           []byte             `json:"description"`
	AuthorID              uuid.UUID          `json:"author_id"`
	AssigneeIds           []uuid.UUID        `json:"assignee_ids"`
	AssigneePositionID    uuid.NullUUID      `json:"assignee_position_id"`
	WatcherIds            []uuid.UUID        `json:"watcher_ids"`
	DueDate               pgtype.Timestamptz `json:"due_date"`
	Priority              string             `json:"priority"`
	LabelIds              []uuid.UUID        `json:"label_ids"`
	Checklist             []byte             `json:"checklist"`
	Attachments           []byte             `json:"attachments"`
	Source                []byte             `json:"source"`
	LinkedArticleIds      []uuid.UUID        `json:"linked_article_ids"`
	Recurrence            []byte             `json:"recurrence"`
	RecurrenceGeneratedAt pgtype.Timestamptz `json:"recurrence_generated_at"`
	CompletedAt           pgtype.Timestamptz `json:"completed_at"`
	DueSoonSentAt         pgtype.Timestamptz `json:"due_soon_sent_at"`
	CreatedAt             time.Time          `json:"created_at"`
	UpdatedAt             time.Time          `json:"updated_at"`
}